# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Warrior Prayers API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listMyGroups
      tags: [groups]
      summary: List all groups the current user belongs to
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupSummary"
        "401":
          description: Unauthorized
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new prayer group
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "401":
          description: Unauthorized
        "403":
          description: Not a group member
        "404":
          description: Group not found
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group settings (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupSettingsInput"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
        "404":
          description: Group not found
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List all active members of a group
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMemberDetail"
        "401":
          description: Unauthorized
        "403":
          description: Not a group member
  /groups/{groupId}/members/{memberId}:
    patch:
      operationId: updateGroupMember
      tags: [groups]
      summary: Update a member's role (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleInput"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMemberDetail"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
        "404":
          description: Member not found
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from a group (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
        "404":
          description: Member not found
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    GroupSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        churchName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        verse:
          type: ["string", "null"]
        memberCount:
          type: integer
        myRole:
          type: string
          enum: [admin, moderator, member]
        createdAt:
          type: string
      required:
        - id
        - name
        - memberCount
        - myRole
        - createdAt
    GroupDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        churchName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        verse:
          type: ["string", "null"]
        hidePrayerPersonNames:
          type: boolean
        allowCustomCategories:
          type: boolean
        allowComments:
          type: boolean
        allowAnonymousRequests:
          type: boolean
        adminsCanViewAnonymousAuthors:
          type: boolean
        memberCount:
          type: integer
        myRole:
          type: string
          enum: [admin, moderator, member]
        createdByUserId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - hidePrayerPersonNames
        - allowCustomCategories
        - allowComments
        - allowAnonymousRequests
        - adminsCanViewAnonymousAuthors
        - memberCount
        - myRole
        - createdByUserId
        - createdAt
        - updatedAt
    GroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        churchName:
          type: string
        city:
          type: string
        imageUrl:
          type: string
        verse:
          type: string
        hidePrayerPersonNames:
          type: boolean
        allowCustomCategories:
          type: boolean
        allowComments:
          type: boolean
        allowAnonymousRequests:
          type: boolean
        adminsCanViewAnonymousAuthors:
          type: boolean
      required:
        - name
    GroupSettingsInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        churchName:
          type: string
        city:
          type: string
        imageUrl:
          type: string
        verse:
          type: string
        hidePrayerPersonNames:
          type: boolean
        allowCustomCategories:
          type: boolean
        allowComments:
          type: boolean
        allowAnonymousRequests:
          type: boolean
        adminsCanViewAnonymousAuthors:
          type: boolean
    GroupMemberDetail:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        groupId:
          type: string
        role:
          type: string
          enum: [admin, moderator, member]
        status:
          type: string
        joinedAt:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
      required:
        - id
        - userId
        - groupId
        - role
        - status
        - joinedAt
        - email
    MemberRoleInput:
      type: object
      properties:
        role:
          type: string
          enum: [admin, moderator, member]
      required:
        - role
...